s3 as a separate module
