+++ asto-s3/src/main/java/com/artipie/asto/s3/S3Storage.java
@@ -39,7 +39,7 @@ import software.amazon.awssdk.services.s3.model.S3Object;
 /**
  * Storage that holds data in S3 storage.
  *
- * @since 0.15
+ * @since 0.1
  * @todo #87:60min Do not await abort to complete if save() failed.
  *  In case uploading content fails inside {@link S3Storage#save(Key, Content)} method
  *  we are doing abort() for multipart upload.
