+++ asto-s3/src/main/java/com/artipie/asto/s3/package-info.java
@@ -4,9 +4,8 @@
  */
 
 /**
- * Tests for S3 storage related classes.
+ * Implementation of storage that holds data in S3.
  *
- * @since 0.15
+ * @since 0.1
  */
 package com.artipie.asto.s3;
-
