+++ asto-s3/src/test/java/com/artipie/asto/package-info.java
@@ -1,0 +1,12 @@
+/*
+ * The MIT License (MIT) Copyright (c) 2020-2022 artipie.com
+ * https://github.com/artipie/asto/LICENSE.txt
+ */
+
+/**
+ * S3 storage tests.
+ *
+ * @since 0.1
+ */
+package com.artipie.asto;
+
